fix: fix celebrate middleware in auth router

+++ src/controllers/EconomicActivityController.js
@@ -1,12 +1,22 @@
-const { EconomicActivity } = require('../models');
+const { EconomicActivity, User } = require('../models');
+const { getJWTBody } = require('../functions/auth/getJWTBody');
 
 module.exports = {
   async getActivities (request, response) {
     try {
-      let getActivities = await EconomicActivity.findAll(request.body);
+      let getToken = request.headers['authorization'];
+      let getId = await getJWTBody(getToken);
+      let getUser = await User.findOne({ raw: true, where: { id: getId, is_super_user: true } });
 
-      response.status(200).json({ body: getActivities });
+      if (getUser == null || getUser == undefined) {
+        response.status(401).json({ error: 'Unauthorized' });
 
+      } else {
+        let getActivities = await EconomicActivity.findAll(request.body);
+
+        response.status(200).json({ body: getActivities });
+      };
+      
     } catch (error) {
       response.status(500).json({ error: error });
     };
@@ -14,17 +24,26 @@ module.exports = {
 
   async registerActivity (request, response) {
     try {
-      let verifyActivity = await EconomicActivity.findOne({
-        where: {name: request.body.name}
-      });
+      let getToken = request.headers['authorization'];
+      let getId = await getJWTBody(getToken);
+      let getUser = await User.findOne({ raw: true, where: { id: getId, is_super_user: true } });
 
-      if (verifyActivity == null || verifyActivity == undefined) {
-        await EconomicActivity.create(request.body);
-        response.status(201).json({ message: 'Activity created' });
+      if (getUser == null || getUser == undefined) {
+        response.status(401).json({ error: 'Unauthorized' });
 
       } else {
+        let verifyActivity = await EconomicActivity.findOne({
+          where: { name: request.body.name }
+        });
+  
+        if (verifyActivity == null || verifyActivity == undefined) {
+          await EconomicActivity.create(request.body);
+          response.status(201).json({ message: 'Activity created' });
+
+        } else {
         response.status(401).json({ error: 'Unauthorized' });
       };
+    };
 
     } catch (error) {
       response.status(500).json({ error: error });
@@ -33,12 +52,19 @@ module.exports = {
 
   async editActivity (request, response) {
     try {
+      let getToken = request.headers['authorization'];
+      let getId = await getJWTBody(getToken);
+      let getUser = await User.findOne({ raw: true, where: { id: getId, is_super_user: true } });
 
-      await EconomicActivity.update(request.body, {
-        where: {id: request.body.id}
-      })
+      if (getUser == null || getUser == undefined) {
+        response.status(401).json({ error: 'Unauthorized' });
 
+      } else {
+        await EconomicActivity.update(request.body, {
+          where: { id: request.body.id }
+        });
       response.status(200).json({ message: 'Activity edited' });
+    };
 
     } catch (error) {
       response.status(500).json({ error: error });
@@ -47,16 +73,23 @@ module.exports = {
 
   async deleteActivity (request, response) {
     try {
-      
-      await EconomicActivity.destroy({
-        where: {id: request.body.id}
-      });
+      let getToken = request.headers['authorization'];
+      let getId = await getJWTBody(getToken);
+      let getUser = await User.findOne({ raw: true, where: { id: getId, is_super_user: true } });
 
+      if (getUser == null || getUser == undefined) {
+        response.status(401).json({ error: 'Unauthorized' });
+
+      } else {
+        await EconomicActivity.destroy({
+          where: {id: request.body.id}
+        });
       response.status(200).json({ message: 'Activity deleted' });
+    };
 
     } catch (error) {
       response.status(500).json({ error: error });
     };
-  }
+  },
 
-}
+};
